+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+- **New**
+  - Added support for `Wall of Stone`
+- **Update**
+  - Fixed issue where `damage-rolls` were being counted for summon sources
+
 ## 1.24.0
 
 - **New**